// src/ui/navigation/navigation.types.ts
import { IsActiveMatchOptions, Params, QueryParamsHandling } from '@angular/router';

export interface Navigation {
	compact?: NavigationItem[];
	default?: NavigationItem[];
	futuristic?: NavigationItem[];
	horizontal?: NavigationItem[];
}

export interface NavigationItem {
	id?: string;
	title?: string;
	subtitle?: string;
	type: 'aside' | 'basic' | 'collapsable' | 'divider' | 'group' | 'spacer';
	hidden?: (item: NavigationItem) => boolean;
	active?: boolean;
	disabled?: boolean;
	tooltip?: string;
	link?: string;
	preserveSubPathByRegex?: RegExp;
	fragment?: string;
	preserveFragment?: boolean;
	queryParams?: Params | null;
	queryParamsHandling?: QueryParamsHandling | null;
	externalLink?: boolean;
	target?: '_blank' | '_self' | '_parent' | '_top' | string;
	exactMatch?: boolean;
	isActiveMatchOptions?: IsActiveMatchOptions;
	function?: (item: NavigationItem) => void;
	classes?: {
		title?: string;
		subtitle?: string;
		icon?: string;
		wrapper?: string;
	};
	icon?: string;
	badge?: {
		title?: string;
		classes?: string;
	};
	children?: NavigationItem[];
	meta?: any;
}